file rename and prep
