+++ css/pacman.css
@@ -1,4 +1,5 @@
 body {
+    background-color: black;
     font-family: Calibri;
     text-align: center;
 }
