+++ html/indexO.html
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <title>PacMan: The Game</title>
+        <meta charset="utf-8">
+        <meta lang="en-us">
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+        <link rel="stylesheet" href="../css/pacman.css">
+    </head>
+    <body>
+        <div id="game-container-id" class="game-container">
+            <div>
+                <p>
+                    Use cursor keys to move player
+                </p>
+            </div>
+            <div class="game-map">
+                <div id="tiles" class="layer"></div>
+                <div id="sprites" class="layer">
+                    <canvas id="canvas" width="224" height="248"></canvas>
+                </div>
+            </div>
+            <img src="../Nursery/spritesheet.png" id="spritesheet" class="hidden">
+        </div>
+
+        <script type="module" src="../source/engine.js"></script>
+    </body>
+</html>
